Introduce ActionDispatch::HostAuthorization for DNS rebinding attack prevention
The idea is simple: DNS rebinding is a pretty common attach and it lets
attackers do a Remote Code Execution in a newly created Rails apps. The
code is executed by Web Console in our case, but the attack is universal
across any web development framework and languages.

A simple way to prevent it is white-listing the hosts a requests is
allowed to be sent to. This has to be configured to any app upgrading to
Rails 6, though, so we have to think of the best way to configure it.